ENH: Add clearer error messages for TLS connection issues
* For both client and server, drive handshake to completion at association establishment, reporting on TLS connection errors.
* For both client and server, provide best effort checking for mismatch where local doesn't support TLS but remote has TLS enabled
