+++ ul/Cargo.toml
@@ -17,6 +17,7 @@ bytes = "1.11.1"
 cfg-if = "1.0.3"
 dicom-encoding = { path = "../encoding/", version = "0.9.1" }
 dicom-transfer-syntax-registry = { path = "../transfer-syntax-registry/", version = "0.9.1", default-features = false }
+dicom-app-common = { version = "0.9", path = "../app-common", optional = true }
 snafu = "0.9"
 tracing = "0.1.34"
 
@@ -64,7 +65,7 @@ rustls = "0.23.31"
 
 [features]
 async = ["dep:tokio"]
-sync-tls = ["dep:rustls"]
+sync-tls = ["dep:rustls", "dep:dicom-app-common"]
 async-tls=["async", "sync-tls", "dep:tokio-rustls"]
 default = []
 full = ["async-tls"]
