+++ ul/src/association/client.rs
@@ -28,6 +28,7 @@ use crate::{
     },
 };
 use snafu::{ResultExt, ensure};
+use tracing::{debug, error};
 
 use super::{Result, uid::trim_uid};
 
@@ -89,13 +90,15 @@ where
 {
     use std::convert::TryFrom;
 
-    let socket = tcp_connection(ae_address, opts)?;
+    let mut socket = tcp_connection(ae_address, opts)?;
     let server_name = rustls::pki_types::ServerName::try_from(server_name.to_string())
         .context(super::InvalidServerNameSnafu)?;
 
-    let conn = rustls::ClientConnection::new(tls_config.clone(), server_name)
+    let mut conn = rustls::ClientConnection::new(tls_config.clone(), server_name)
         .context(super::TlsConnectionSnafu)?;
-
+    let (read, wrote) = conn.complete_io(&mut socket)
+        .context(super::TlsHandshakeSnafu)?;
+    debug!(read_bytes=read, write_bytes=wrote, "Client handshake completed");
     Ok(rustls::StreamOwned::new(conn, socket))
 }
 
@@ -921,7 +924,44 @@ impl<'a> ClientAssociationOptions<'a> {
         let mut buf = BytesMut::with_capacity(
             (self.max_pdu_length.min(LARGE_PDU_SIZE) + PDU_HEADER_SIZE) as usize,
         );
-        let resp = read_pdu_from_wire(&mut socket, &mut buf, self.max_pdu_length, self.strict)?;
+        let resp = read_pdu_from_wire(&mut socket, &mut buf, self.max_pdu_length, self.strict)
+            .map_err(|e| {
+                #[cfg(feature = "sync-tls")]
+                {
+                    let options = dicom_app_common::TlsOptions::default();
+                    let config =  match options.client_config().map(std::sync::Arc::new){
+                        Ok(c) => c,
+                        Err(_) => return e
+                    };
+                    let server_name = match rustls::pki_types::ServerName::try_from("localhost"){
+                        Ok(n) => n,
+                        Err(_) => return e
+                    };
+                    let mut cursor = std::io::Cursor::new(buf.to_vec());
+                    let mut conn = rustls::ClientConnection::new(config, server_name).unwrap();
+                    if conn.read_tls(&mut cursor).is_err(){
+                        return e
+                    }
+                    match conn.process_new_packets(){
+                        Ok(_) => {
+                            // This shouldn't be possible.  If we're here this
+                            // means the server responded with a valid TLS
+                            // state, but since we didn't originally send a TLS
+                            // ClientHello, that shouldn't be possible
+                            error!("Recieved TLS response to non-TLS request!");
+                            super::TlsNotSupportedSnafu.build()
+                        }
+                        Err(e) => {
+                            error!(err=?e, "Server expects TLS, client sent plaintext");
+                            super::TlsNotSupportedSnafu.build()
+                        }
+                    }
+                }
+                #[cfg(not(feature = "sync-tls"))]
+                {
+                    e
+                }
+            })?;
         let negotiated_options = self.process_a_association_resp(resp, &pc_proposed);
         match negotiated_options {
             Err(e) => {
@@ -1315,7 +1355,19 @@ where
     let tls_stream = connector
         .connect(domain, tcp_stream)
         .await
-        .context(crate::association::ConnectSnafu)?;
+        .map_err(|e| {
+            // This is how tokio_rustls represents a handshake error
+            if let std::io::ErrorKind::UnexpectedEof = e.kind() {
+                return super::Error::TlsHandshake{
+                    source: e,
+                    backtrace: std::backtrace::Backtrace::capture()
+                }
+            }
+            super::Error::Connect{
+                source: e,
+                backtrace: std::backtrace::Backtrace::capture()
+            }
+        })?;
     Ok(tls_stream)
 }
 
@@ -1400,8 +1452,45 @@ impl<'a> ClientAssociationOptions<'a> {
             )
             .await
         })
-        .await?;
-        let negotiated_options = self.process_a_association_resp(resp, &pc_proposed);
+        .await;
+        let pdu = match resp  {
+            Ok(pdu) => pdu,
+            Err(e) => {
+                // Attempt to check if the failure was because the client sent a TLS stream
+                #[cfg(feature = "async-tls")]
+                {
+                    let options = dicom_app_common::TlsOptions::default();
+                    let config =  match options.client_config().map(std::sync::Arc::new){
+                        Ok(c) => c,
+                        Err(_) => return Err(e)
+                    };
+                    let server_name = match rustls::pki_types::ServerName::try_from("localhost"){
+                        Ok(n) => n,
+                        Err(_) => return Err(e)
+                    };
+                    let mut cursor = std::io::Cursor::new(read_buffer.to_vec());
+                    let mut conn = rustls::ClientConnection::new(config, server_name).unwrap();
+                    if conn.read_tls(&mut cursor).is_err(){
+                        return Err(e)
+                    }
+                    match conn.process_new_packets(){
+                        Ok(_) => {
+                            // This shouldn't be possible.  If we're here this
+                            // means the server responded with a valid TLS
+                            // state, but since we didn't originally send a TLS
+                            // ClientHello, that shouldn't be possible
+                            error!("Recieved TLS response to non-TLS request!");
+                            return super::TlsNotSupportedSnafu.fail();
+                        }
+                        Err(e) => {
+                            error!(err=?e, "Server expects TLS, client sent plaintext");
+                            return super::TlsNotSupportedSnafu.fail();
+                        }
+                    }
+                }
+            }
+        };
+        let negotiated_options = self.process_a_association_resp(pdu, &pc_proposed);
         match negotiated_options {
             Err(e) => {
                 // abort connection
